+++ packages/plugin-text-color/web/src/viewer.ts
@@ -3,6 +3,11 @@ import { TEXT_COLOR_TYPE, TEXT_HIGHLIGHT_TYPE } from './types';
 import { DEFAULTS } from './constants';
 import type { ViewerPluginCreator } from 'wix-rich-content-common';
 export { TEXT_COLOR_TYPE, TEXT_HIGHLIGHT_TYPE };
+import {
+  RICOS_TEXT_COLOR_TYPE,
+  RICOS_TEXT_HIGHLIGHT_TYPE,
+  getDynamicInlineStyleMapper,
+} from 'wix-rich-content-common';
 
 export const pluginTextColor: ViewerPluginCreator<TextColorPluginViewerConfig> = config => {
   return {
@@ -17,3 +22,8 @@ export const pluginTextHighlight: ViewerPluginCreator<TextHighlightPluginViewerC
     type: TEXT_HIGHLIGHT_TYPE,
   };
 };
+
+export const textColorInlineStyleMapper = getDynamicInlineStyleMapper(RICOS_TEXT_COLOR_TYPE);
+
+export const textHighlightInlineStyleMapper =
+  getDynamicInlineStyleMapper(RICOS_TEXT_HIGHLIGHT_TYPE);